+++ instrumentation/navigation-compose-nav2/src/test/kotlin/io/opentelemetry/android/instrumentation/navigation/compose/nav2/ComposeNav2CollectorTest.kt
@@ -57,6 +57,8 @@ class ComposeNav2CollectorTest {
         assertThat(spans[0].attributes.get(NavigationConstants.NAVIGATION_DESTINATION_TYPE_KEY)).isEqualTo("compose_route")
         assertThat(spans[0].attributes.get(NavigationConstants.NAVIGATION_TRANSITION_TYPE_KEY)).isEqualTo("push")
         assertThat(spans[0].attributes.get(NavigationConstants.NAVIGATION_TRIGGER_KEY)).isEqualTo("unknown")
+        assertThat(spans[0].attributes.get(NavigationConstants.NAVIGATION_STACK_DEPTH_BEFORE_KEY)).isEqualTo(0L)
+        assertThat(spans[0].attributes.get(NavigationConstants.NAVIGATION_STACK_DEPTH_AFTER_KEY)).isEqualTo(1L)
     }
 
     @Test